// Client/src/utils/api/schemas/scoreboardDataTableColumn.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Turnierverwaltung
 * OpenAPI spec version: 1.0.0
 */
import type { ScoreboardDataTableColumnAlignment } from "./scoreboardDataTableColumnAlignment";
import type { ScoreboardDataTableColumnIWidth } from "./scoreboardDataTableColumnIWidth";

export interface ScoreboardDataTableColumn {
  name: string;
  width: ScoreboardDataTableColumnIWidth;
  alignment: ScoreboardDataTableColumnAlignment;
}