// Client/src/utils/api/schemas/scoreboardDataTableColumnAlignment.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Turnierverwaltung
 * OpenAPI spec version: 1.0.0
 */

export type ScoreboardDataTableColumnAlignment =
  (typeof ScoreboardDataTableColumnAlignment)[keyof typeof ScoreboardDataTableColumnAlignment];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ScoreboardDataTableColumnAlignment = {
  Left: "Left",
  Center: "Center",
  Right: "Right",
} as const;